+++ src/monopoly/banks/dbs/dbs.py
@@ -3,6 +3,7 @@ from re import compile as regex
 
 from monopoly.config import StatementConfig
 from monopoly.constants import (
+    ISO8601,
     BankNames,
     CreditTransactionPatterns,
     DebitTransactionPatterns,
@@ -17,23 +18,33 @@ logger = logging.getLogger(__name__)
 
 
 class Dbs(BankBase):
-    credit_config = StatementConfig(
+    name = BankNames.DBS
+
+    credit = StatementConfig(
         statement_type=EntryType.CREDIT,
-        bank_name=BankNames.DBS,
-        statement_date_pattern=regex(r"(\d{2}\s[A-Za-z]{3}\s\d{4})"),
+        statement_date_pattern=ISO8601.DD_MMM_YYYY,
         multiline_transactions=False,
         header_pattern=regex(r"(DATE.*DESCRIPTION.*AMOUNT)"),
         transaction_pattern=CreditTransactionPatterns.DBS,
         prev_balance_pattern=StatementBalancePatterns.DBS,
     )
 
-    debit_config = StatementConfig(
+    debit = StatementConfig(
         statement_type=EntryType.DEBIT,
-        bank_name=BankNames.DBS,
-        statement_date_pattern=regex(r"(\d{2}\s[A-Za-z]{3}\s\d{4})"),
+        statement_date_pattern=ISO8601.DD_MMM_YYYY,
         multiline_transactions=True,
         header_pattern=regex(r"(Withdrawal.*Deposit.*Balance)"),
         transaction_pattern=DebitTransactionPatterns.DBS,
+        transaction_bound=170,
+    )
+
+    consolidated = StatementConfig(
+        statement_type=EntryType.DEBIT,
+        statement_date_pattern=regex(rf"Details as at {ISO8601.DD_MMM_YYYY}"),
+        multiline_transactions=True,
+        header_pattern=regex(r"(\s+Date\s+Description\s+Withdrawal \(-\).*)"),
+        transaction_pattern=DebitTransactionPatterns.DBS_POSB_CONSOLIDATED,
+        transaction_bound=100,
     )
 
     identifiers = [
@@ -43,4 +54,4 @@ class Dbs(BankBase):
         ],
     ]
 
-    statement_configs = [debit_config, credit_config]
+    statement_configs = [credit, consolidated, debit]