+++ src/monopoly/banks/hsbc/hsbc.py
@@ -3,6 +3,7 @@ from re import compile as regex
 
 from monopoly.config import PdfConfig, StatementConfig
 from monopoly.constants import (
+    ISO8601,
     BankNames,
     CreditTransactionPatterns,
     EntryType,
@@ -16,10 +17,11 @@ logger = logging.getLogger(__name__)
 
 
 class Hsbc(BankBase):
-    credit_config = StatementConfig(
+    name = BankNames.HSBC
+
+    credit = StatementConfig(
         statement_type=EntryType.CREDIT,
-        bank_name=BankNames.HSBC,
-        statement_date_pattern=regex(r"Statement From .* to (\d{2}\s[A-Z]{3}\s\d{4})"),
+        statement_date_pattern=regex(rf"Statement From .* to {ISO8601.DD_MMM_YYYY}"),
         header_pattern=regex(r"(DATE.*DESCRIPTION.*AMOUNT)"),
         prev_balance_pattern=StatementBalancePatterns.HSBC,
         transaction_pattern=CreditTransactionPatterns.HSBC,
@@ -48,4 +50,4 @@ class Hsbc(BankBase):
 
     identifiers = [email_statement_identifier, web_and_mobile_statement_identifier]
 
-    statement_configs = [credit_config]
+    statement_configs = [credit]