+++ src/monopoly/banks/ocbc/ocbc.py
@@ -3,6 +3,7 @@ from re import compile as regex
 
 from monopoly.config import StatementConfig
 from monopoly.constants import (
+    ISO8601,
     BankNames,
     CreditTransactionPatterns,
     DebitTransactionPatterns,
@@ -17,22 +18,23 @@ logger = logging.getLogger(__name__)
 
 
 class Ocbc(BankBase):
-    credit_config = StatementConfig(
+    name = BankNames.OCBC
+
+    credit = StatementConfig(
         statement_type=EntryType.CREDIT,
-        bank_name=BankNames.OCBC,
-        statement_date_pattern=regex(r"(\d{2}\-\d{2}\-\d{4})"),
+        statement_date_pattern=ISO8601.DD_MM_YYYY,
         header_pattern=regex(r"(TRANSACTION DATE.*DESCRIPTION.*AMOUNT)"),
         prev_balance_pattern=StatementBalancePatterns.OCBC,
         transaction_pattern=CreditTransactionPatterns.OCBC,
     )
 
-    debit_config = StatementConfig(
+    debit = StatementConfig(
         statement_type=EntryType.DEBIT,
-        bank_name=BankNames.OCBC,
-        statement_date_pattern=regex(r"TO\s(\d+\s[A-Za-z]{3}\s\d{4})"),
+        statement_date_pattern=regex(rf"\s{ISO8601.DD_MMM_YYYY}$"),
         header_pattern=regex(r"(Withdrawal.*Deposit.*Balance)"),
         transaction_pattern=DebitTransactionPatterns.OCBC,
         multiline_transactions=True,
+        transaction_bound=170,
     )
 
     identifiers = [
@@ -44,4 +46,4 @@ class Ocbc(BankBase):
         ],
     ]
 
-    statement_configs = [debit_config, credit_config]
+    statement_configs = [credit, debit]