+++ src/monopoly/banks/standard_chartered/standard_chartered.py
@@ -3,6 +3,7 @@ from re import compile as regex
 
 from monopoly.config import StatementConfig
 from monopoly.constants import (
+    ISO8601,
     BankNames,
     CreditTransactionPatterns,
     EntryType,
@@ -16,10 +17,11 @@ logger = logging.getLogger(__name__)
 
 
 class StandardChartered(BankBase):
-    credit_config = StatementConfig(
+    name = BankNames.STANDARD_CHARTERED
+
+    credit = StatementConfig(
         statement_type=EntryType.CREDIT,
-        bank_name=BankNames.STANDARD_CHARTERED,
-        statement_date_pattern=regex(r"(\d{2}\s\w+\s\d{4})"),
+        statement_date_pattern=regex(rf": {ISO8601.DD_MMM_YYYY}$"),
         header_pattern=regex(r"(Transaction.*Posting.*Amount)"),
         prev_balance_pattern=StatementBalancePatterns.STANDARD_CHARTERED,
         transaction_pattern=CreditTransactionPatterns.STANDARD_CHARTERED,
@@ -35,4 +37,4 @@ class StandardChartered(BankBase):
         ]
     ]
 
-    statement_configs = [credit_config]
+    statement_configs = [credit]