+++ src/monopoly/banks/zkb/zkb.py
@@ -2,8 +2,7 @@ import logging
 from re import compile as regex
 
 from monopoly.config import StatementConfig
-from monopoly.constants import BankNames, DebitTransactionPatterns, EntryType
-from monopoly.constants.date import ISO8601
+from monopoly.constants import ISO8601, BankNames, DebitTransactionPatterns, EntryType
 from monopoly.identifiers import MetadataIdentifier
 
 from ..base import BankBase
@@ -12,9 +11,10 @@ logger = logging.getLogger(__name__)
 
 
 class ZurcherKantonalBank(BankBase):
-    debit_config = StatementConfig(
+    name = BankNames.ZKB
+
+    debit = StatementConfig(
         statement_type=EntryType.DEBIT,
-        bank_name=BankNames.ZKB,
         statement_date_pattern=regex(rf"Balance as of: ({ISO8601.DD_MM_YYYY})"),
         header_pattern=regex(r"(Date.*Booking text.*Debit CHF.*Credit CHF)"),
         transaction_pattern=DebitTransactionPatterns.ZKB,
@@ -32,4 +32,4 @@ class ZurcherKantonalBank(BankBase):
             )
         ]
     ]
-    statement_configs = [debit_config]
+    statement_configs = [debit]