+++ tests/integration/banks/test_banks_generic_debit.py
@@ -6,6 +6,7 @@ from test_utils.skip import skip_if_encrypted
 from test_utils.transactions import get_transactions_as_dict, read_transactions_from_csv
 
 from monopoly.banks import BankBase, Dbs, Maybank, Ocbc
+from monopoly.generic import GenericBank
 from monopoly.pdf import PdfDocument, PdfParser
 from monopoly.pipeline import Pipeline
 from monopoly.statements import DebitStatement
@@ -17,28 +18,22 @@ test_cases = [
 ]
 
 
-@pytest.fixture
-def no_banks(monkeypatch):
-    monkeypatch.setattr("monopoly.banks.banks", [])
-
-
 @skip_if_encrypted
 @pytest.mark.parametrize(
     "bank, expected_debit_sum, expected_credit_sum, statement_date",
     test_cases,
 )
-@pytest.mark.usefixtures("no_banks")
 def test_bank_debit_statements(
     bank: BankBase,
     expected_debit_sum: float,
     expected_credit_sum: float,
     statement_date: datetime,
 ):
-    bank_name = bank.debit_config.bank_name
-    test_directory = Path(__file__).parent / bank_name / "debit"
+    test_directory = Path(__file__).parent / bank.name / "debit"
 
     document = PdfDocument(test_directory / "input.pdf")
     parser = PdfParser(bank, document)
+    parser.bank = GenericBank
     pipeline = Pipeline(parser)
     statement: DebitStatement = pipeline.extract()